+++ Assets/Scripts/UI/HealthBar.cs
@@ -16,13 +16,13 @@ namespace RY
             slider = GetComponent<Slider>();
         }
 
-        public void SetMaxHealth(int maxHealth)
+        public void SetMaxHealth(float maxHealth)
         {
             slider.maxValue = maxHealth;
             slider.value = maxHealth;
         }
 
-        public void SetCurrentHealth(int currentHealth)
+        public void SetCurrentHealth(float currentHealth)
         {
             slider.value = currentHealth;
         }
